reorganize observable map
