+++ src/main/kotlin/org/jitsi/nlj/transform/node/incoming/RtxHandler.kt
@@ -24,7 +24,7 @@ import org.jitsi.nlj.rtp.RtxPacket
 import org.jitsi.nlj.rtp.SsrcAssociationType
 import org.jitsi.nlj.stats.NodeStatsBlock
 import org.jitsi.nlj.transform.node.TransformerNode
-import org.jitsi.nlj.util.MapEventValueFilterHandler
+import org.jitsi.nlj.util.observable.map.MapEventValueFilterHandler
 import org.jitsi.nlj.util.ReadOnlyStreamInformationStore
 import org.jitsi.nlj.util.cdebug
 import org.jitsi.nlj.util.cerror
